# openapi.yaml
openapi: 3.0.3
info:
  description: Express boilerplate for building RESTful APIs
  version: 2.0.0
  title: express-rest-boilerplate
...
servers:
  - description: development
    url: http://localhost:3000
  - description: production
    url: https://node-express-rest-boilerplate.herokuapp.com
paths:
...
  /v1/auth/send-password-reset:
    post:
      tags: [auth]
      summary: Send password reset link
      operationId: sendPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: User's email
              example: { 'email': 'your@email.com' }
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: User's email
                password:
                  type: string
                  description: User's new password
                  minLength: 8
                resetToken:
                  type: string
                  description: Reset token sent to user's email
              example: { 'email': 'your@email.com', 'password': '********', 'accessToken': 'abc' }
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: JWT Authentication
...
  responses:
    BadRequest:
      description: 'Bad Request: Some parameters may contain invalid values'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { 'code': 'BadRequest', 'message': 'Something is required' }
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
...